+++ src/ga/Population.java
@@ -9,22 +9,18 @@ public class Population {
 
     public Population() {
         chromosomes = new ArrayList<>();
-
     }
 
     public void addChromosome(Chromosome chromosome) {
         chromosomes.add(chromosome);
-
     }
 
     public List<Chromosome> getChromosomes() {
         return chromosomes;
-
     }
 
     public int size() {
         return chromosomes.size();
-
     }
 
 }