+++ src/ga/constraints/MaxDailyLectureConstraint.java
@@ -1,10 +1,10 @@
 package ga.constraints;
 
-import config.SchedulingConfig;
-import ga.Chromosome;
 import ga.EvaluationContext;
+import model.Day;
 import model.Placement;
 import model.Subject;
+import util.PlacementIndex;
 
 import java.util.HashMap;
 import java.util.List;
@@ -14,7 +14,6 @@ public class MaxDailyLectureConstraint implements Constraint {
 
     private final int maxDailyLimit;
 
-    // default constructor
     public MaxDailyLectureConstraint() {
         this(2);
     }
@@ -27,28 +26,26 @@ public class MaxDailyLectureConstraint implements Constraint {
     public int evaluate(EvaluationContext context) {
         int penalty = 0;
 
-        Placement[][] grid = context.getGrid();
+        PlacementIndex index = context.getIndex();
+        for (Day day : Day.values()) {
 
-        for (int day = 0; day < SchedulingConfig.WORKING_DAYS; day++) {
+            List<Placement> placements = index.getByDay(day);
             Map<Subject, Integer> frequency = new HashMap<>();
 
-            for (int period = 0; period < SchedulingConfig.PERIODS_PER_DAY; period++) {
-                Placement placement = grid[day][period];
+            for (Placement placement : placements) {
+                Subject subject = placement.getSession().getSubject();
 
-                if (placement != null) {
-                    Subject subject = placement.getSession().getSubject();
-                    frequency.put(subject, frequency.getOrDefault(subject, 0) + 1);
-                }
+                if (subject.isLab())
+                    continue;
+                frequency.put(subject, frequency.getOrDefault(subject, 0) + 1);
             }
-            // checks for one working day each loop
-            for (Map.Entry<Subject, Integer> entry : frequency.entrySet()) {
-                int count = entry.getValue();
+
+            for (int count : frequency.values()) {
                 if (count > maxDailyLimit) {
                     penalty += (count - maxDailyLimit) * 10;
                 }
             }
         }
-
         return penalty;
     }
 }