+++ src/ga/constraints/SpreadConstraint.java
@@ -1,6 +1,5 @@
 package ga.constraints;
 
-import ga.Chromosome;
 import ga.EvaluationContext;
 import model.Day;
 import model.Placement;
@@ -9,28 +8,28 @@ import util.PlacementIndex;
 
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 public class SpreadConstraint implements Constraint {
 
     @Override
     public int evaluate(EvaluationContext context) {
-
         int penalty = 0;
         PlacementIndex index = context.getIndex();
 
-        for (Subject subject : index.getBySubject().keySet()) {
+        for (Map.Entry<Subject, List<Placement>> entry : index.getBySubject().entrySet()) {
+            Subject subject = entry.getKey();
 
             if (subject.isLab())
                 continue;
 
-            List<Placement> placements = index.getBySubject(subject);
+            List<Placement> placements = entry.getValue();
             Set<Day> days = new HashSet<>();
-
             for (Placement placement : placements) {
                 days.add(placement.getSlot().getDay());
-
             }
+
             int requiredDays = subject.getCredits();
             int actualDays = days.size();