+++ src/model/Session.java
@@ -27,7 +27,6 @@ public class Session {
 
     @Override
     public String toString() {
-
         return subject.getCode() + "(" + duration + ")";
     }
 }