+++ src/model/Timetable.java
@@ -3,19 +3,22 @@ package model;
 import java.util.ArrayList;
 import java.util.List;
 
-//new : Placement[][] -> List<Placement>
+@Deprecated
 public class Timetable {
 
     private final List<Placement> placements;
 
+    @Deprecated
     public Timetable() {
         placements = new ArrayList<>();
     }
 
+    @Deprecated
     public void addPlacement(Placement placement) {
         placements.add(placement);
     }
 
+    @Deprecated
     public List<Placement> getPlacements() {
         return placements;
     }